chore: clear mypy+ruff backlog, gate CI on lint and types
- pragmatic mypy config (drop blanket strict, scope-out deprecated contrib,
  stub scipy/pandas/networkx, ignore stubless numba/pyarrow/_rust_core)
- mypy 0 across 107 files, ruff 0, format clean
- fix real bugs surfaced by the sweep (risk.py _formatting NameError)
- flip CI ruff/ruff-format/mypy steps to blocking gates
- refresh stale docs (maturin single backend, 78 rust tests)

+++ src/prefgraph/core/results/diagnostics.py
@@ -637,7 +637,7 @@ class BradleyTerryResult(ResultDisplayMixin, ResultPlotMixin):
         s_j = self.scores[item_j]
         exp_i = np.exp(s_i - max(s_i, s_j))
         exp_j = np.exp(s_j - max(s_i, s_j))
-        return exp_i / (exp_i + exp_j)
+        return float(exp_i / (exp_i + exp_j))
 
     def score(self) -> float:
         """Return scikit-learn style score in [0, 1]. Higher is better.
